requirements.txt
Django==2.1.4
djangorestframework==3.9.0
lxml==4.2.5
requests==2.21.0
